Changed location of doget
